organize the functions for usertouser chat
